Fixes for robustness
Increase blend to 1e-2 (more experimentation is probably needed). Do one more iteration of refinement (this should almost certainly be conditional for performance). And, most importantly, fix circular arc math.

Not bad but this should be considered an intermediate state. More robustness is needed for zero derivatives, and the performance impact should be mitigated.