More or less working evolute approximation
Not yet doing subdivision, as I want to get the non-subdivided version right.

This is holding t's constant on the approximation and doing a Newton step on the source curve. The other way around might be just as good, and would save from having to compute the derivative of the evolute (currently done numerically).

It's not very stable. Number of iterations is fixed at 5, but at least it's protected against error increasing.